chore(appium): use different ports fo config files
