+++ config/wdio.windows.chatA.conf.ts
@@ -1,8 +1,9 @@
+import config from "./wdio.shared.windows.appium.conf";
+import { join } from "path";
+
+
 const fsp = require("fs").promises;
 const mkdirp = require("mkdirp");
-
-import config from "./wdio.shared.mac.appium.conf";
-import { join } from "path";
 
 // ============
 // Specs
@@ -17,21 +18,22 @@ config.specs = [join(process.cwd(), "./tests/specs/13-chats-userA.spec.ts")];
 config.capabilities = [
   {
     // The defaults you need to have in your config
-    platformName: "mac",
+    platformName: "windows",
+    "appium:deviceName": "WindowsPC",
     // For W3C the appium capabilities need to have an extension prefix
     // http://appium.io/docs/en/writing-running-appium/caps/
     // This is `appium:` for all Appium Capabilities which can be found here
-    "appium:automationName": "mac2",
-    // @ts-ignore
-    "appium:bundleId": "im.satellite.uplink",
-    "appium:serverStartupTimeout": 240000,
+    "appium:automationName": "windows",
+    "appium:app": join(process.cwd(), "\\apps\\ui.exe"),
+    "ms:waitForAppLaunch": 30,
+    "appium:systemPort": 4725,
   },
 ];
 
 config.afterTest = async function (test, describe, { error }) {
   if (error) {
     let imageFile = await driver.takeScreenshot();
-    let imageFolder = join(process.cwd(), "./test-results/macos", test.parent);
+    let imageFolder = join(process.cwd(), "./test-results/windows", test.parent);
     await mkdirp(imageFolder);
     await fsp.writeFile(
       imageFolder + "/" + test.title + " - Failed.png",
